Improved SQL builder: - ID column is now of datatype long and auto increment - Removed ID adaption within builder framework (bad performance) - Added In condition for ID

+++ ojcms-sqlbuilder/src/main/java/de/adito/ojcms/sqlbuilder/AbstractSQLStatement.java
@@ -26,7 +26,7 @@ public abstract class AbstractSQLStatement<RESULT, STATEMENT extends AbstractSQL
   protected final AbstractSQLBuilder builder;
   protected final IDatabasePlatform databasePlatform;
   protected final IValueSerializer serializer;
-  protected final IColumnIdentification<Integer> idColumnIdentification;
+  protected final IColumnIdentification<Long> idColumnIdentification;
   private String tableName;
 
   /**
@@ -45,7 +45,7 @@ public abstract class AbstractSQLStatement<RESULT, STATEMENT extends AbstractSQL
     builder = pBuilder;
     databasePlatform = pPlatform;
     serializer = pSerializer;
-    idColumnIdentification = IColumnIdentification.of(pIdColumnName.toUpperCase(), Integer.class);
+    idColumnIdentification = IColumnIdentification.of(pIdColumnName.toUpperCase(), Long.class);
   }
 
   /**
